requirements.txt
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
prompt-toolkit==3.0.41
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
